move progressbar to own directory
- add doc strings
- split out decor into its own file
- remove dependency on `events`
