+++ locale/en-us.yaml
@@ -1642,3 +1642,9 @@ err_activate_platfrom_alternate_branches:
     Type [ACTIONABLE]`state branch switch <NAME>`[/RESET] to switch to a different branch.
 err_set_default_branch:
   other: Could not update project field with branch, please manually set the `branch={{.V0}}` query parameter on the project field in your activestate.yaml.
+artifact_progress_step_download:
+  other: Downloading
+artifact_progress_step_unpack:
+  other: Unpacking
+artifact_progress_step_install:
+  other: Installing
